GridHelpers

Documentation for GridHelpers.

GridHelpers.Cell

GridHelpers.bilinear_weights

GridHelpers.interpolate_bilinear

GridHelpers.Cell — Type

Cell defined by four corner points around coordinates (x, y). (x, y) should be floating-point indices associated to 1-based arrays, i.e. the rounded value of x and y should be an integer-valued grid index.

Iteration order is defined as follows: bottom left, bottom right, top left, top right.

GridHelpers.bilinear_weights — Method

Return bilinear weights from all four corners, in cell iteration order. Extracting bilinear weights with this function may be useful when they are to be combined with other weights.

Note

If you want to interpolate a value directly, use interpolate_bilinear which will be much faster.

GridHelpers.interpolate_bilinear — Function

Perform a bilinear interpolation of A at location (x, y).

If the cell is provided,